pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "humming"
version = "0.1.0"
description = "Quantization GEMM Kernel"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "torch",
    "pyelftools",
    "pycuda"
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
"*" = ["*.cuh"]

[tool.ruff]
line-length = 100
